+++ src/models/payment.ts
@@ -2,8 +2,8 @@ import { DataTypes, Model, Optional } from 'sequelize'
 import sequelize from '@config/database'
 import { RentalAttributes } from '@type/rentals'
 import { v4 as uuid } from 'uuid'
-import Customer from '@models/customers'
-import Location from '@models/locations'
+import Customer from '@models/customer'
+import Location from '@models/location'
 import User from '@models/user'
 
 class Rental
