+++ src/models/product.ts
@@ -2,7 +2,7 @@ import { DataTypes, Model, Optional } from 'sequelize'
 import sequelize from '@config/database'
 import { ProductAttributes } from '@type/products'
 import { v4 as uuid } from 'uuid'
-import Category from '@models/categories'
+import Category from '@models/category'
 
 class Product
   extends Model<ProductAttributes, Optional<ProductAttributes, 'id'>>
