+++ src/routes/resource.ts
